backend/requirements.txt
flask
flask-cors
google-generativeai
psycopg2-binary
gunicorn
flask-bcrypt
flask-jwt-extended